@@ -3,66 +3,81 @@ name: appoint
 description: "A new Flutter project."
 publish_to: 'none'
 
-version: 1.0.0+1
+version: 1.0.0+0
 
 environment:
-  sdk: ">=3.3.0 <4.0.0"
-  flutter: ">=3.19.0"
+  sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  intl: ^0.20.2
-  firebase_core: ^2.21.1
-  cloud_firestore: ^4.6.1
-  firebase_auth: ^4.6.1
-  firebase_messaging: ^14.7.10
-  cloud_functions: ^4.7.5
-  firebase_analytics: ^10.4.0
-  firebase_storage: ^11.5.1
-  firebase_crashlytics: ^3.5.7
-  firebase_remote_config: ^4.3.3
-  firebase_app_check: ^0.2.1+8
+  intl: ^0.19.0
+  firebase_core: ^3.15.0
+  cloud_firestore: ^5.6.10
+  firebase_auth: ^5.6.1
+  firebase_messaging: ^15.2.8
+  cloud_functions: ^5.6.0
+  firebase_analytics: ^11.5.1
+  firebase_storage: ^12.4.8
+  firebase_crashlytics: ^4.3.8
+  firebase_remote_config: ^5.4.6
+  firebase_app_check: ^0.3.2+8
+  firebase_performance: ^0.10.1+8
   flutter_stripe: ^11.5.0
   flutter_riverpod: 2.6.1
   freezed_annotation: ^2.4.4
-  json_annotation: ^4.9.0
-  equatable: ^2.0.5
-  flutter_bloc: ^8.1.4
-  google_mobile_ads: ^4.0.0
-  http: ^1.1.0
-  url_launcher: ^6.2.4
-  googleapis: ^12.0.0
-  googleapis_auth: ^1.4.1
-  flutter_web_auth_2: ^2.0.4
-  flutter_secure_storage: ^9.0.0
-  uni_links: ^0.5.1
-  share_plus: ^7.2.1
-  file_picker: ^6.1.1
+  google_sign_in: ^6.1.4
+  sign_in_with_apple: ^6.1.2
+  file_picker: ^8.1.4
+  permission_handler: ^11.3.1
   image_picker: ^1.0.7
-  google_maps_flutter: 2.6.1
-  geolocator: 12.0.0
-  flutter_dotenv: ^5.1.0
-  mailer: ^6.0.1
-  fl_chart: 0.71.0
-  go_router: ^13.2.0
-  shadcn_ui: 0.15.2
-  tailwind_colors: ^0.3.1
-  syncfusion_flutter_charts: 28.2.6
+  just_audio: ^0.9.40
+  carousel_slider: ^5.0.0
+  flutter_markdown: 0.7.4+1
+  path_provider: ^2.1.4
+  url_launcher: ^6.3.1
+  package_info_plus: ^8.0.3
+  device_info_plus: ^10.1.2
+  crypto: ^3.0.5
+  google_fonts: ^6.2.1
+  pin_code_fields: ^8.0.1
+  fl_chart: ^0.71.0
+  go_router: ^13.2.4
+  syncfusion_flutter_charts: ^25.2.7
+  shared_preferences: ^2.2.2
+  connectivity_plus: ^5.0.2
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
 
+  webview_flutter: ^4.7.0
+  uuid: ^4.1.0
+  async: ^2.10.0
+  dio: ^5.4.3+1
+  stripe_platform_interface: ^11.5.0
+  flutter_local_notifications: ^17.2.2
 dev_dependencies:
+  integration_test:
+    sdk: flutter
+  mocktail: ^1.0.4
+  fake_cloud_firestore: ^3.1.0
+  hive_test: ^1.0.1
+  build_runner: ^2.4.10
+  hive_generator: ^2.0.1
+  freezed: ^2.5.2
+  json_serializable: ^6.8.0
+  flutter_lints: ^3.0.2
+
   flutter_test:
     sdk: flutter
-  build_runner: ^2.4.7
-  freezed: ^2.5.2
-  json_serializable: ^6.7.1
-  fake_cloud_firestore: ^2.4.0
-  firebase_auth_mocks: ^0.13.0
-  mockito: ^5.4.2
-  test: ^1.24.9
-
+  firebase_auth_mocks: ^0.14.1
+  firebase_auth_platform_interface: ^7.0.0
+  cloud_firestore_platform_interface: ^6.6.10
+  firebase_crashlytics_platform_interface: ^3.8.8
+  firebase_core_platform_interface: ^5.0.0
+  plugin_platform_interface: ^2.1.8
+  mockito: ^5.0.17
 flutter:
   uses-material-design: true
   generate: true
@@ -77,8 +92,3 @@ flutter:
         pluginClass: FilePickerPlugin
       ios:
         pluginClass: FilePickerPlugin
-
-dependency_overrides:
-  intl: ^0.20.2
-  fl_chart:
-    path: packages/fl_chart_stub
