@@ -1,5 +1,5 @@
 import 'package:appoint/features/messaging/models/message.dart';
-import 'package:appoint/features/messaging/services/messaging_service.dart';
+// import 'package:appoint/features/messaging/services/messaging_service.dart'; // Unused
 import 'package:appoint/l10n/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
