@@ -1,11 +1,12 @@
 import 'package:appoint/features/messaging/models/message.dart';
 import 'package:flutter/material.dart';
+// import 'package:url_launcher/url_launcher.dart'; // Unused
 
 class MessageBubble extends StatelessWidget {
   const MessageBubble({
+    super.key,
     required this.message,
     required this.isMe,
-    super.key,
     this.onTap,
   });
 
@@ -14,73 +15,69 @@ class MessageBubble extends StatelessWidget {
   final VoidCallback? onTap;
 
   @override
-  Widget build(BuildContext context) => Align(
-        alignment: isMe ? Alignment.centerRight : Alignment.centerLeft,
-        child: Container(
-          margin: EdgeInsets.only(
-            left: isMe ? 64 : 8,
-            right: isMe ? 8 : 64,
-          ),
-          child: Column(
-            crossAxisAlignment:
-                isMe ? CrossAxisAlignment.end : CrossAxisAlignment.start,
-            children: [
-              // Message bubble
-              GestureDetector(
-                onTap: onTap,
-                child: Container(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-                  decoration: BoxDecoration(
-                    color: isMe
-                        ? Theme.of(context).primaryColor
-                        : Colors.grey[200],
-                    borderRadius: BorderRadius.circular(20).copyWith(
-                      bottomLeft: isMe
-                          ? const Radius.circular(20)
-                          : const Radius.circular(4),
-                      bottomRight: isMe
-                          ? const Radius.circular(4)
-                          : const Radius.circular(20),
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: isMe ? Alignment.centerRight : Alignment.centerLeft,
+      child: Container(
+        margin: EdgeInsets.only(
+          left: isMe ? 64 : 8,
+          right: isMe ? 8 : 64,
+        ),
+        child: Column(
+          crossAxisAlignment: isMe ? CrossAxisAlignment.end : CrossAxisAlignment.start,
+          children: [
+            // Message bubble
+            GestureDetector(
+              onTap: onTap,
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+                decoration: BoxDecoration(
+                  color: isMe 
+                      ? Theme.of(context).primaryColor 
+                      : Colors.grey[200],
+                  borderRadius: BorderRadius.circular(20).copyWith(
+                    bottomLeft: isMe ? const Radius.circular(20) : const Radius.circular(4),
+                    bottomRight: isMe ? const Radius.circular(4) : const Radius.circular(20),
+                  ),
+                ),
+                child: _buildMessageContent(),
+              ),
+            ),
+            
+            const SizedBox(height: 4),
+            
+            // Message metadata
+            Padding(
+              padding: EdgeInsets.only(
+                left: isMe ? 0 : 16,
+                right: isMe ? 16 : 0,
+              ),
+              child: Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Text(
+                    _formatTime(message.timestamp),
+                    style: TextStyle(
+                      fontSize: 12,
+                      color: Colors.grey[500],
                     ),
                   ),
-                  child: _buildMessageContent(),
-                ),
-              ),
-
-              const SizedBox(height: 4),
-
-              // Message metadata
-              Padding(
-                padding: EdgeInsets.only(
-                  left: isMe ? 0 : 16,
-                  right: isMe ? 16 : 0,
-                ),
-                child: Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Text(
-                      _formatTime(message.timestamp),
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: Colors.grey[500],
-                      ),
+                  if (isMe) ...[
+                    const SizedBox(width: 4),
+                    Icon(
+                      message.isRead ? Icons.done_all : Icons.done,
+                      size: 16,
+                      color: message.isRead ? Colors.blue : Colors.grey[500],
                     ),
-                    if (isMe) ...[
-                      const SizedBox(width: 4),
-                      Icon(
-                        message.isRead ? Icons.done_all : Icons.done,
-                        size: 16,
-                        color: message.isRead ? Colors.blue : Colors.grey[500],
-                      ),
-                    ],
                   ],
-                ),
+                ],
               ),
-            ],
-          ),
+            ),
+          ],
         ),
-      );
+      ),
+    );
+  }
 
   Widget _buildMessageContent() {
     switch (message.type) {
@@ -92,22 +89,22 @@ class MessageBubble extends StatelessWidget {
             fontSize: 16,
           ),
         );
-
+      
       case MessageType.image:
         return _buildImageMessage();
-
+      
       case MessageType.video:
         return _buildVideoMessage();
-
+      
       case MessageType.audio:
         return _buildAudioMessage();
-
+      
       case MessageType.file:
         return _buildFileMessage();
-
+      
       case MessageType.location:
         return _buildLocationMessage();
-
+      
       case MessageType.system:
         return _buildSystemMessage();
     }
@@ -140,12 +137,14 @@ class MessageBubble extends StatelessWidget {
             width: 200,
             height: 200,
             fit: BoxFit.cover,
-            errorBuilder: (context, error, stackTrace) => Container(
-              width: 200,
-              height: 200,
-              color: Colors.grey[300],
-              child: const Icon(Icons.broken_image, size: 48),
-            ),
+            errorBuilder: (context, error, stackTrace) {
+              return Container(
+                width: 200,
+                height: 200,
+                color: Colors.grey[300],
+                child: const Icon(Icons.broken_image, size: 48),
+              );
+            },
           ),
         ),
       ],
@@ -188,18 +187,19 @@ class MessageBubble extends StatelessWidget {
                   width: 200,
                   height: 150,
                   fit: BoxFit.cover,
-                  errorBuilder: (context, error, stackTrace) => Container(
-                    color: Colors.grey[800],
-                    child: const Icon(Icons.video_library,
-                        color: Colors.white, size: 48),
-                  ),
+                  errorBuilder: (context, error, stackTrace) {
+                    return Container(
+                      color: Colors.grey[800],
+                      child: const Icon(Icons.video_library, color: Colors.white, size: 48),
+                    );
+                  },
                 ),
               ),
               Center(
                 child: Container(
                   padding: const EdgeInsets.all(8),
                   decoration: BoxDecoration(
-                    color: Colors.black.withValues(alpha: 0.5),
+                    color: Colors.black.withOpacity(0.5),
                     shape: BoxShape.circle,
                   ),
                   child: const Icon(
@@ -216,33 +216,34 @@ class MessageBubble extends StatelessWidget {
     );
   }
 
-  Widget _buildAudioMessage() => Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(
-            Icons.play_arrow,
+  Widget _buildAudioMessage() {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Icon(
+          Icons.play_arrow,
+          color: isMe ? Colors.white : Colors.black87,
+        ),
+        const SizedBox(width: 8),
+        Container(
+          width: 100,
+          height: 4,
+          decoration: BoxDecoration(
+            color: isMe ? Colors.white.withOpacity(0.3) : Colors.grey[400],
+            borderRadius: BorderRadius.circular(2),
+          ),
+        ),
+        const SizedBox(width: 8),
+        Text(
+          '0:30',
+          style: TextStyle(
             color: isMe ? Colors.white : Colors.black87,
+            fontSize: 12,
           ),
-          const SizedBox(width: 8),
-          Container(
-            width: 100,
-            height: 4,
-            decoration: BoxDecoration(
-              color:
-                  isMe ? Colors.white.withValues(alpha: 0.3) : Colors.grey[400],
-              borderRadius: BorderRadius.circular(2),
-            ),
-          ),
-          const SizedBox(width: 8),
-          Text(
-            '0:30',
-            style: TextStyle(
-              color: isMe ? Colors.white : Colors.black87,
-              fontSize: 12,
-            ),
-          ),
-        ],
-      );
+        ),
+      ],
+    );
+  }
 
   Widget _buildFileMessage() {
     if (message.attachments?.isEmpty ?? true) {
@@ -267,8 +268,7 @@ class MessageBubble extends StatelessWidget {
         Container(
           padding: const EdgeInsets.all(12),
           decoration: BoxDecoration(
-            color:
-                isMe ? Colors.white.withValues(alpha: 0.2) : Colors.grey[100],
+            color: isMe ? Colors.white.withOpacity(0.2) : Colors.grey[100],
             borderRadius: BorderRadius.circular(8),
           ),
           child: Row(
@@ -293,9 +293,7 @@ class MessageBubble extends StatelessWidget {
                     Text(
                       _formatFileSize(attachment.size!),
                       style: TextStyle(
-                        color: isMe
-                            ? Colors.white.withValues(alpha: 0.7)
-                            : Colors.grey[600],
+                        color: isMe ? Colors.white.withOpacity(0.7) : Colors.grey[600],
                         fontSize: 12,
                       ),
                     ),
@@ -308,78 +306,81 @@ class MessageBubble extends StatelessWidget {
     );
   }
 
-  Widget _buildLocationMessage() => Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          if (message.content.isNotEmpty)
-            Padding(
-              padding: const EdgeInsets.only(bottom: 8),
-              child: Text(
-                message.content,
-                style: TextStyle(
-                  color: isMe ? Colors.white : Colors.black87,
-                  fontSize: 16,
-                ),
+  Widget _buildLocationMessage() {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        if (message.content.isNotEmpty)
+          Padding(
+            padding: const EdgeInsets.only(bottom: 8),
+            child: Text(
+              message.content,
+              style: TextStyle(
+                color: isMe ? Colors.white : Colors.black87,
+                fontSize: 16,
               ),
             ),
-          Container(
-            width: 200,
-            height: 150,
-            decoration: BoxDecoration(
-              color: Colors.grey[300],
-              borderRadius: BorderRadius.circular(12),
-            ),
-            child: Stack(
-              children: [
-                ClipRRect(
-                  borderRadius: BorderRadius.circular(12),
-                  child: Container(
-                    color: Colors.grey[300],
-                    child: const Icon(Icons.map, size: 48, color: Colors.grey),
-                  ),
-                ),
-                Positioned(
-                  bottom: 8,
-                  left: 8,
-                  right: 8,
-                  child: Container(
-                    padding:
-                        const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    decoration: BoxDecoration(
-                      color: Colors.black.withValues(alpha: 0.7),
-                      borderRadius: BorderRadius.circular(4),
-                    ),
-                    child: const Text(
-                      'Tap to open in Maps',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 12,
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
           ),
-        ],
-      );
+        Container(
+          width: 200,
+          height: 150,
+          decoration: BoxDecoration(
+            color: Colors.grey[300],
+            borderRadius: BorderRadius.circular(12),
+          ),
+          child: Stack(
+            children: [
+              ClipRRect(
+                borderRadius: BorderRadius.circular(12),
+                child: Container(
+                  color: Colors.grey[300],
+                  child: const Icon(Icons.map, size: 48, color: Colors.grey),
+                ),
+              ),
+              Positioned(
+                bottom: 8,
+                left: 8,
+                right: 8,
+                child: Container(
+                  padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
+                  decoration: BoxDecoration(
+                    color: Colors.black.withOpacity(0.7),
+                    borderRadius: BorderRadius.circular(4),
+                  ),
+                  child: const Text(
+                    'Tap to open in Maps',
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 12,
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
 
-  Widget _buildSystemMessage() => Container(
-        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-        decoration: BoxDecoration(
-          color: Colors.grey[100],
-          borderRadius: BorderRadius.circular(12),
+  Widget _buildSystemMessage() {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+      decoration: BoxDecoration(
+        color: Colors.grey[100],
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Text(
+        message.content,
+        style: TextStyle(
+          color: Colors.grey[600],
+          fontSize: 14,
+          fontStyle: FontStyle.italic,
         ),
-        child: Text(
-          message.content,
-          style: TextStyle(
-            color: Colors.grey[600],
-            fontSize: 14,
-            fontStyle: FontStyle.italic,
-          ),
-          textAlign: TextAlign.center,
-        ),
-      );
+        textAlign: TextAlign.center,
+      ),
+    );
+  }
 
   IconData _getFileIcon(AttachmentType type) {
     switch (type) {
@@ -399,8 +400,7 @@ class MessageBubble extends StatelessWidget {
   String _formatFileSize(int bytes) {
     if (bytes < 1024) return '$bytes B';
     if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
-    if (bytes < 1024 * 1024 * 1024)
-      return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';
+    if (bytes < 1024 * 1024 * 1024) return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';
     return '${(bytes / (1024 * 1024 * 1024)).toStringAsFixed(1)} GB';
   }
 
@@ -417,4 +417,4 @@ class MessageBubble extends StatelessWidget {
       return '${time.day}/${time.month}/${time.year}';
     }
   }
-}
+} 
