@@ -1,7 +1,7 @@
 import 'package:appoint/l10n/app_localizations.dart';
 import 'package:appoint/models/notification_settings.dart';
 import 'package:appoint/providers/auth_provider.dart';
-import 'package:appoint/providers/fcm_token_provider.dart';
+// import 'package:appoint/providers/notification_provider.dart'; // Unused
 import 'package:appoint/providers/user_settings_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -11,34 +11,37 @@ class NotificationSettingsScreen extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, final WidgetRef ref) {
-    final l10n = AppLocalizations.of(context)!;
-    final tokenState = ref.watch(fcmTokenProvider);
-    final settingsAsync = ref.watch(notificationSettingsProvider);
+    l10n = AppLocalizations.of(context)!;
+    tokenAsync = ref.watch(fcmTokenProvider);
+    settingsAsync = ref.watch(notificationSettingsProvider);
     return Scaffold(
       appBar: AppBar(title: Text(l10n.notificationSettings)),
       body: settingsAsync.when(
-        data: (settings) => Padding(
-          padding: const EdgeInsets.all(16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              SwitchListTile(
-                title: Text(l10n.enableNotifications),
-                value: settings.push,
-                onChanged: (v) async {
-                  final uid = ref.read(authProvider).currentUser?.uid;
-                  if (uid == null) return;
-                  final newSettings = NotificationSettings(push: v);
-                  await ref
-                      .read(userSettingsServiceProvider)
-                      .updateSettings(uid, newSettings);
-                },
+        data: (settings) => tokenAsync.when(
+            data: (token) => Padding(
+                padding: const EdgeInsets.all(16),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    SwitchListTile(
+                      title: Text(l10n.enableNotifications),
+                      value: settings.push,
+                      onChanged: (v) async {
+                        uid = ref.read(authProvider).currentUser?.uid;
+                        if (uid == null) return;
+                        newSettings = NotificationSettings(push: v);
+                        await ref
+                            .read(userSettingsServiceProvider)
+                            .updateSettings(uid, newSettings);
+                      },
+                    ),
+                    if (token != null) SelectableText(l10n.fcmToken(token)),
+                  ],
+                ),
               ),
-              if (tokenState.token != null)
-                SelectableText(l10n.fcmToken(tokenState.token!)),
-            ],
+            loading: () => const Center(child: CircularProgressIndicator()),
+            error: (e, final _) => Center(child: Text('Error: $e')),
           ),
-        ),
         loading: () => const Center(child: CircularProgressIndicator()),
         error: (e, final _) => Center(child: Text('Error: $e')),
       ),
