@@ -1,39 +1,29 @@
 import 'package:flutter/material.dart';
 
-/// Enhanced onboarding screen with improved user experience
+class OnboardingScreen extends StatelessWidget {
+  const OnboardingScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Onboarding')),
+      body: const Center(
+        child: Text('Onboarding Screen'),
+      ),
+    );
+  }
+}
+
 class EnhancedOnboardingScreen extends StatelessWidget {
   const EnhancedOnboardingScreen({super.key});
 
   @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(
-          title: const Text('Enhanced Onboarding'),
-        ),
-        body: const Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Icon(
-                Icons.rocket_launch,
-                size: 80,
-                color: Colors.blue,
-              ),
-              SizedBox(height: 20),
-              Text(
-                'Welcome to APP-OINT!',
-                style: TextStyle(
-                  fontSize: 24,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              SizedBox(height: 10),
-              Text(
-                'Enhanced onboarding experience coming soon...',
-                style: TextStyle(fontSize: 16),
-                textAlign: TextAlign.center,
-              ),
-            ],
-          ),
-        ),
-      );
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Enhanced Onboarding')),
+      body: const Center(
+        child: Text('Enhanced Onboarding Screen'),
+      ),
+    );
+  }
 }
