@@ -1,18 +1,19 @@
 import 'dart:async';
-import 'package:uni_links/uni_links.dart';
-import 'package:flutter/material.dart';
+
+import 'package:appoint/config/environment_config.dart';
+import 'package:appoint/services/whatsapp_share_service.dart';
 import 'package:flutter/foundation.dart' show kIsWeb;
-
-import 'whatsapp_share_service.dart';
-
+import 'package:flutter/material.dart';
 class CustomDeepLinkService {
-  late final WhatsAppShareService _whatsappService;
   CustomDeepLinkService({WhatsAppShareService? whatsappShareService}) {
     _whatsappService =
         whatsappShareService ?? WhatsAppShareService(deepLinkService: this);
   }
+  late final WhatsAppShareService _whatsappService;
   StreamSubscription? _linkSubscription;
   StreamSubscription? _initialLinkSubscription;
+  // GlobalKey is kept for future navigation features.
+  // ignore: unused_field
   GlobalKey<NavigatorState>? _navigatorKey;
 
   /// Set the navigator key for navigation
@@ -28,7 +29,7 @@ class CustomDeepLinkService {
     }
     try {
       // Handle initial link if app was opened from a link
-      final initialUri = await getInitialUri();
+      initialUri = await getInitialUri();
       if (initialUri != null) {
         await _handleDeepLink(initialUri);
       }
@@ -41,7 +42,7 @@ class CustomDeepLinkService {
           }
         },
         onError: (error) {
-          print('Deep link error: $error');
+          // Removed debug print: debugPrint('Deep link error: $error');
         },
       );
 
@@ -53,11 +54,11 @@ class CustomDeepLinkService {
           }
         },
         onError: (error) {
-          print('App link error: $error');
+          // Removed debug print: debugPrint('App link error: $error');
         },
       );
-    } catch (e) {
-      print('Error initializing deep links: $e');
+    } catch (e) {e) {
+      // Removed debug print: debugPrint('Error initializing deep links: $e');
     }
   }
 
@@ -69,7 +70,7 @@ class CustomDeepLinkService {
     // ensure proper configuration for each platform.
     /*
     try {
-      print('Handling deep link: $uri');
+      // Removed debug print: debugPrint('Handling deep link: $uri');
 
       // Let WhatsApp service handle the link first for analytics
       await _whatsappService.handleDeepLink(uri);
@@ -105,17 +106,18 @@ class CustomDeepLinkService {
             }
             break;
           default:
-            print('Unknown deep link path: ${pathSegments[0]}');
+            // Removed debug print: debugPrint('Unknown deep link path: ${pathSegments[0]}');
         }
       }
-    } catch (e) {
-      print('Error handling deep link: $e');
+    } catch (e) {e) {
+      // Removed debug print: debugPrint('Error handling deep link: $e');
     }
     */
   }
 
-  /// Navigate to meeting details screen
-  // ignore: unused_element
+  // Navigate helper methods are kept for potential future use. They are
+  // commented out to silence analysis warnings.
+  /*
   Future<void> _navigateToMeeting(
     String meetingId,
     String? creatorId,
@@ -133,12 +135,10 @@ class CustomDeepLinkService {
         },
       );
     } else {
-      print('Navigator key not set, cannot navigate to meeting: $meetingId');
+      // Removed debug print: debugPrint('Navigator key not set, cannot navigate to meeting: \$meetingId');
     }
   }
 
-  /// Navigate to invite details screen
-  // ignore: unused_element
   Future<void> _navigateToInvite(String inviteId) async {
     if (_navigatorKey?.currentState != null) {
       _navigatorKey!.currentState!.pushNamed(
@@ -146,12 +146,10 @@ class CustomDeepLinkService {
         arguments: {'inviteId': inviteId},
       );
     } else {
-      print('Navigator key not set, cannot navigate to invite: $inviteId');
+      // Removed debug print: debugPrint('Navigator key not set, cannot navigate to invite: \$inviteId');
     }
   }
 
-  /// Navigate to booking details screen
-  // ignore: unused_element
   Future<void> _navigateToBooking(String bookingId) async {
     if (_navigatorKey?.currentState != null) {
       _navigatorKey!.currentState!.pushNamed(
@@ -159,18 +157,22 @@ class CustomDeepLinkService {
         arguments: {'bookingId': bookingId},
       );
     } else {
-      print('Navigator key not set, cannot navigate to booking: $bookingId');
+      // Removed debug print: debugPrint('Navigator key not set, cannot navigate to booking: \$bookingId');
     }
   }
+  */
 
   /// Create a deep link for a meeting using custom URL scheme
   Future<String> createMeetingLink({
-    required String meetingId,
-    required String creatorId,
-    String? contextId,
-    String? groupId,
+    required final String meetingId,
+    required final String creatorId,
+    final String? contextId,
+    final String? groupId,
   }) async {
     try {
+      // Load base URL from environment configuration
+      const baseUrl = EnvironmentConfig.deepLinkBaseUrl;
+
       // Create a custom URL scheme link
       final queryParams = <String, String>{
         'creatorId': creatorId,
@@ -188,8 +190,7 @@ class CustomDeepLinkService {
           .map((e) => '${e.key}=${Uri.encodeComponent(e.value)}')
           .join('&');
 
-      final link =
-          'https://app-oint-core.web.app/meeting/$meetingId?$queryString';
+      final link = '$baseUrl/meeting/$meetingId?$queryString';
 
       // Store share analytics
       await _whatsappService.logShareAnalytics(
@@ -200,10 +201,11 @@ class CustomDeepLinkService {
       );
 
       return link;
-    } catch (e) {
-      print('Error creating meeting link: $e');
+    } catch (e) {e) {
+      // Removed debug print: debugPrint('Error creating meeting link: $e');
       // Fallback to simple URL
-      return 'https://app-oint-core.web.app/meeting/$meetingId?creatorId=$creatorId${contextId != null ? '&contextId=$contextId' : ''}';
+      const baseUrl = EnvironmentConfig.deepLinkBaseUrl;
+      return '$baseUrl/meeting/$meetingId?creatorId=$creatorId${contextId != null ? '&contextId=$contextId' : ''}';
     }
   }
 
@@ -218,8 +220,8 @@ class CustomDeepLinkService {
       // For now, return the long URL
       // In production, you would call a URL shortener API
       return longUrl;
-    } catch (e) {
-      print('Error creating short link: $e');
+    } catch (e) {e) {
+      // Removed debug print: debugPrint('Error creating short link: $e');
       return longUrl;
     }
   }
