@@ -1,37 +1,51 @@
 import 'dart:convert';
 
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:appoint/config/google_config.dart';
+import 'package:encrypt/encrypt.dart' as encrypt;
 import 'package:flutter_web_auth_2/flutter_web_auth_2.dart';
-import 'package:googleapis/calendar/v3.dart';
 import 'package:googleapis_auth/googleapis_auth.dart';
 import 'package:http/http.dart' as http;
-import 'package:encrypt/encrypt.dart' as encrypt;
-
-import '../config/google_config.dart';
 
 class GoogleCalendarService {
+
+  GoogleCalendarService({FlutterSecureStorage? storage})
+      : _storage = storage ?? const FlutterSecureStorage();
   static const _credentialKey = 'google_calendar_credentials';
   static const _encryptionKeyKey = 'google_calendar_encryption_key';
 
   final FlutterSecureStorage _storage;
   AutoRefreshingAuthClient? _client;
 
-  GoogleCalendarService({FlutterSecureStorage? storage})
-      : _storage = storage ?? const FlutterSecureStorage();
-
   Future<encrypt.Key> _getEncryptionKey() async {
-    final existing = await _storage.read(key: _encryptionKeyKey);
+    existing = await _storage.read(key: _encryptionKeyKey);
     if (existing != null) {
       return encrypt.Key.fromBase64(existing);
     }
-    final key = encrypt.Key.fromSecureRandom(32);
+    key = encrypt.Key.fromSecureRandom(32);
     await _storage.write(key: _encryptionKeyKey, value: key.base64);
     return key;
   }
 
+  Future<String> _encryptData(String plainText) async {
+    key = await _getEncryptionKey();
+    iv = encrypt.IV.fromSecureRandom(16);
+    encrypter = encrypt.Encrypter(encrypt.AES(key));
+    encrypted = encrypter.encrypt(plainText, iv: iv);
+    return jsonEncode({'iv': iv.base64, 'data': encrypted.base64});
+  }
+
+  Future<String> _decryptData(String payload) async {
+    key = await _getEncryptionKey();
+    map = jsonDecode(payload) as Map<String, dynamic>;
+    iv = encrypt.IV.fromBase64(map['iv'] as String);
+    final data = map['data'] as String;
+    encrypter = encrypt.Encrypter(encrypt.AES(key));
+    return encrypter.decrypt64(data, iv: iv);
+  }
+
   Future<void> signInWithGoogleCalendar() async {
     // If credentials already stored, just load them.
-    final creds = await _loadCredentials();
+    creds = await _loadCredentials();
     if (creds != null) {
       _setClientFromCredentials(creds);
       return;
@@ -51,7 +65,7 @@ class GoogleCalendarService {
       callbackUrlScheme: Uri.parse(GoogleConfig.redirectUri).scheme,
     );
 
-    final code = Uri.parse(result).queryParameters['code'];
+    code = Uri.parse(result).queryParameters['code'];
     if (code == null || code.isEmpty) {
       throw Exception('Authorization code not found');
     }
@@ -71,7 +85,7 @@ class GoogleCalendarService {
       throw Exception('Token exchange failed: ${tokenResp.body}');
     }
 
-    final Map<String, dynamic> data = jsonDecode(tokenResp.body);
+    Map<String, dynamic> data = jsonDecode(tokenResp.body);
     final accessToken = AccessToken(
       'Bearer',
       data['access_token'] as String,
@@ -90,22 +104,22 @@ class GoogleCalendarService {
   }
 
   Future<List<CalendarListEntry>> getCalendars() async {
-    final client = await _getClient();
-    final api = CalendarApi(client);
-    final list = await api.calendarList.list();
+    client = await _getClient();
+    api = CalendarApi(client);
+    list = await api.calendarList.list();
     return list.items ?? <CalendarListEntry>[];
   }
 
   Future<void> createEvent(
-    String calendarId, {
-    required String summary,
-    required DateTime start,
-    required DateTime end,
-    String? description,
+    final String calendarId, {
+    required final String summary,
+    required final DateTime start,
+    required final DateTime end,
+    final String? description,
   }) async {
-    final client = await _getClient();
-    final api = CalendarApi(client);
-    final event = Event()
+    client = await _getClient();
+    api = CalendarApi(client);
+    event = Event()
       ..summary = summary
       ..description = description
       ..start = (EventDateTime()..dateTime = start.toUtc())
@@ -116,45 +130,44 @@ class GoogleCalendarService {
 
   Future<AutoRefreshingAuthClient> _getClient() async {
     if (_client != null) return _client!;
-    final creds = await _loadCredentials();
+    creds = await _loadCredentials();
     if (creds == null) {
       throw Exception('Not authenticated with Google');
     }
     return _setClientFromCredentials(creds);
   }
 
-  AutoRefreshingAuthClient _setClientFromCredentials(AccessCredentials creds) {
-    final clientId = ClientId(GoogleConfig.clientId, null);
-    final client = autoRefreshingClient(clientId, creds, http.Client());
+  AutoRefreshingAuthClient _setClientFromCredentials(
+      AccessCredentials creds,) {
+    clientId = ClientId(GoogleConfig.clientId);
+    client = autoRefreshingClient(clientId, creds, http.Client());
     client.credentialUpdates.listen(_saveCredentials);
     _client = client;
     return client;
   }
 
   Future<AccessCredentials?> _loadCredentials() async {
-    final stored = await _storage.read(key: _credentialKey);
-    if (stored == null) return null;
+    encrypted = await _storage.read(key: _credentialKey);
+    if (encrypted == null) return null;
     try {
-      final key = await _getEncryptionKey();
-      final wrapper = jsonDecode(stored) as Map<String, dynamic>;
-      final iv = encrypt.IV.fromBase64(wrapper['iv'] as String);
-      final encryptedData = wrapper['data'] as String;
-      final encrypter = encrypt.Encrypter(encrypt.AES(key));
-      final decrypted = encrypter.decrypt64(encryptedData, iv: iv);
-      final data = jsonDecode(decrypted) as Map<String, dynamic>;
+      decrypted = await _decryptData(encrypted);
+      data = jsonDecode(decrypted) as Map<String, dynamic>;
       return AccessCredentials.fromJson(data);
-    } catch (_) {
+    } catch (e) {_) {
       return null;
     }
   }
 
   Future<void> _saveCredentials(AccessCredentials credentials) async {
-    final key = await _getEncryptionKey();
-    final iv = encrypt.IV.fromSecureRandom(16);
-    final encrypter = encrypt.Encrypter(encrypt.AES(key));
-    final jsonData = jsonEncode(credentials.toJson());
-    final encrypted = encrypter.encrypt(jsonData, iv: iv);
-    final wrapper = jsonEncode({'iv': iv.base64, 'data': encrypted.base64});
-    await _storage.write(key: _credentialKey, value: wrapper);
+    jsonData = jsonEncode(credentials.toJson());
+    encrypted = await _encryptData(jsonData);
+    await _storage.write(key: _credentialKey, value: encrypted);
+  }
+
+  Future<void> signOut() async {
+    await _storage.delete(key: _credentialKey);
+    await _storage.delete(key: _encryptionKeyKey);
+    _client?.close();
+    _client = null;
   }
 }
