@@ -1,6 +1,6 @@
 import 'package:appoint/l10n/app_localizations.dart';
 import 'package:appoint/models/appointment.dart';
-import 'package:appoint/providers/whatsapp_share_provider.dart';
+// import 'package:appoint/providers/whatsapp_share_provider.dart'; // Unused
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
